requirements.txt
django==5.2.5
python-dotenv==1.1.1
djangorestframework==3.16.1
psycopg2-binary==2.9.10
psycopg2==2.9.10
faker==37.5.3